Add Parser type for specifying options
The default parser behavior was to error on any unknown flag option.
This introduces a way to allow unknown fields to be silently ignored
rather than returning an error from Parse.

A ConfigFlag method is also added to Parser so returned the config
file flag.Value may use the Parser's options.